// apps/frontend/app/api/v1/osograph/types/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
  [P in K]-?: NonNullable<T[P]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: string; output: string };
  JSON: { input: Record<string, unknown>; output: Record<string, unknown> };
};
// ...
export type DataConnectionConnection = {
  __typename?: "DataConnectionConnection";
  edges: Array<DataConnectionEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type Dataset = {
  __typename?: "Dataset";
  createdAt: Scalars["DateTime"]["output"];
  creator: User;
  creatorId: Scalars["ID"]["output"];
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  /** Whether the current user's organization is subscribed to this dataset. */
  isSubscribed: Scalars["Boolean"]["output"];
  /**
   * The materializations for this dataset. Returns all materializations regardless of their source
   * (data model, static model, or data ingestion).
   */
  materializations: MaterializationConnection;
  name: Scalars["String"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  /**
   * The runs for this dataset. For USER_MODEL datasets, each run is related to individual data models.
   * FOR DATA_INGESTION datasets, each run is related to individual ingestion jobs.
   * For DATA_CONNECTION datasets, runs are not applicable and this field will be an empty list.
   */
  runs: RunConnection;
  tables: TableConnection;
  type: DatasetType;
  typeDefinition: DatasetTypeDefinition;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type DatasetConnection = {
  __typename?: "DatasetConnection";
  edges: Array<DatasetEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type DatasetEdge = {
  __typename?: "DatasetEdge";
  cursor: Scalars["String"]["output"];
  node: Dataset;
};

/**
 * Dataset types
 *
 * * USER_MODEL: Contains user-defined models
 * * DATA_CONNECTION: Derived from a data connector to an external data source
 * * DATA_INGESTION: Derived from data ingestion pipelines
 */
export const DatasetType = {
  DataConnection: "DATA_CONNECTION",
  DataIngestion: "DATA_INGESTION",
  StaticModel: "STATIC_MODEL",
  UserModel: "USER_MODEL",
} as const;

export type DatasetType = (typeof DatasetType)[keyof typeof DatasetType];
export type DatasetTypeDefinition =
  | DataConnectionDefinition
  | DataIngestionDefinition
  | DataModelDefinition
  | StaticModelDefinition;
// ...
export type MaterializationConnection = {
  __typename?: "MaterializationConnection";
  edges: Array<MaterializationEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type NotebookConnection = {
  __typename?: "NotebookConnection";
  edges: Array<NotebookEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type NotebookEdge = {
  __typename?: "NotebookEdge";
  cursor: Scalars["String"]["output"];
  node: Notebook;
};

export type Organization = {
  __typename?: "Organization";
  createdAt: Scalars["DateTime"]["output"];
  dataConnections: DataConnectionConnection;
  datasets: DatasetConnection;
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  members: UserConnection;
  name: Scalars["String"]["output"];
  notebooks: NotebookConnection;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type OrganizationConnection = {
  __typename?: "OrganizationConnection";
  edges: Array<OrganizationEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type PageInfo = {
  __typename?: "PageInfo";
  endCursor?: Maybe<Scalars["String"]["output"]>;
  hasNextPage: Scalars["Boolean"]["output"];
  hasPreviousPage: Scalars["Boolean"]["output"];
  startCursor?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Run = {
  __typename?: "Run";
  dataset?: Maybe<Dataset>;
  datasetId?: Maybe<Scalars["ID"]["output"]>;
  finishedAt?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["ID"]["output"];
  /**
   * The logs urls is returned as a presigned url to a cloud storage location. Use
   * a GET request on this URL to retrieve the logs for the run.
   */
  logsUrl?: Maybe<Scalars["String"]["output"]>;
  metadata?: Maybe<Scalars["JSON"]["output"]>;
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  queuedAt: Scalars["DateTime"]["output"];
  requestedBy?: Maybe<User>;
  runType: RunType;
  startedAt?: Maybe<Scalars["DateTime"]["output"]>;
  status: RunStatus;
  steps: StepConnection;
  triggerType: RunTriggerType;
};

export type RunStepsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type RunConnection = {
  __typename?: "RunConnection";
  edges: Array<RunEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type RunEdge = {
  __typename?: "RunEdge";
  cursor: Scalars["String"]["output"];
  node: Run;
};

export const RunStatus = {
  Canceled: "CANCELED",
  Failed: "FAILED",
  Queued: "QUEUED",
  Running: "RUNNING",
  Success: "SUCCESS",
} as const;

export type RunStatus = (typeof RunStatus)[keyof typeof RunStatus];
/**
 * The scheduler schema defines types and queries related to scheduling data model
 * runs and ingest jobs.
 *
 * Runs are a generic representation of a data task that has been executed. So it's
 * meant to be quite flexible to accommodate different types of runs in the future.
 */
export const RunTriggerType = {
  Manual: "MANUAL",
  Scheduled: "SCHEDULED",
} as const;

export type RunTriggerType =
  (typeof RunTriggerType)[keyof typeof RunTriggerType];
export const RunType = {
  Manual: "MANUAL",
  Scheduled: "SCHEDULED",
} as const;

export type RunType = (typeof RunType)[keyof typeof RunType];
// ...
export type StepConnection = {
  __typename?: "StepConnection";
  edges: Array<StepEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type StepEdge = {
  __typename?: "StepEdge";
  cursor: Scalars["String"]["output"];
  node: Step;
};
// ...
export type TableConnection = {
  __typename?: "TableConnection";
  edges: Array<TableEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type User = {
  __typename?: "User";
  avatarUrl?: Maybe<Scalars["String"]["output"]>;
  email: Scalars["String"]["output"];
  fullName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  organizations: OrganizationConnection;
};

export type UserOrganizationsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type UserConnection = {
  __typename?: "UserConnection";
  edges: Array<UserEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<
  TResult,
  TParent = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
  TArgs = Record<PropertyKey, never>,
> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo,
) => Promise<TResult> | TResult;
// ...
export type RunResolvers<
  ContextType = GraphQLContext,
  ParentType extends ResolversParentTypes["Run"] = ResolversParentTypes["Run"],
> = {
  dataset?: Resolver<Maybe<ResolversTypes["Dataset"]>, ParentType, ContextType>;
  datasetId?: Resolver<Maybe<ResolversTypes["ID"]>, ParentType, ContextType>;
  finishedAt?: Resolver<
    Maybe<ResolversTypes["DateTime"]>,
    ParentType,
    ContextType
  >;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  logsUrl?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  metadata?: Resolver<Maybe<ResolversTypes["JSON"]>, ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  organization?: Resolver<
    ResolversTypes["Organization"],
    ParentType,
    ContextType
  >;
  queuedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  requestedBy?: Resolver<
    Maybe<ResolversTypes["User"]>,
    ParentType,
    ContextType
  >;
  runType?: Resolver<ResolversTypes["RunType"], ParentType, ContextType>;
  startedAt?: Resolver<
    Maybe<ResolversTypes["DateTime"]>,
    ParentType,
    ContextType
  >;
  status?: Resolver<ResolversTypes["RunStatus"], ParentType, ContextType>;
  steps?: Resolver<
    ResolversTypes["StepConnection"],
    ParentType,
    ContextType,
    RequireFields<RunStepsArgs, "first">
  >;
  triggerType?: Resolver<
    ResolversTypes["RunTriggerType"],
    ParentType,
    ContextType
  >;
};